requirements.txt
aprs3==8.0.0.post1
kiss3==8.0.0
gpsdclient==1.3.2
textual==0.41.0